+++ .github/workflows/main.yml
@@ -21,7 +21,7 @@ on:
 
 jobs:
   build_and_test_gcc_old:
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-20.04
     name: Build & Test GCC
     env:
       COMPILER_VERS: 7 8
@@ -76,7 +76,7 @@ jobs:
           done;
 
   build_and_test_clang_old:
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-20.04
     name: Build & Test clang
     env:
       COMPILER_VERS: 7 8 9
