### 2023-05-15 a507a82
Improve channelised data quality/perf with overlaps/smaller FFTs
Add baseline for coherent DM support
