+++ src/CLI/lofar_cli_stokes.c
@@ -1,6 +1,7 @@
 #include "lofar_cli_meta.h"
 
 // Import FFTW3 for channelisation
+#include <math.h>
 #include "fftw3.h"
 
 // Import omp for parallelisaing the channelisation/downsamling operations
@@ -8,6 +9,8 @@
 
 // Constant to define the length of the timing variable array
 #define TIMEARRLEN 7
+
+#define dmPhaseConst 2.41e-10 // Hidden somewhere in Hankins et al, can't find a copy
 
 // Output Stokes Modes
 typedef enum stokes_t {
@@ -30,7 +33,9 @@ void helpMessages() {
 
 }
 
-static void CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_t *header, fftwf_complex *X, fftwf_complex *Y) {
+static void
+CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_t *header, fftwf_complex (*X), fftwf_complex (*Y), fftwf_complex *in1,
+           fftwf_complex (*in2), fftwf_complex (*chirpData)) {
 
 	FREE_NOT_NULL(outConfig);
 	FREE_NOT_NULL(config);
@@ -39,29 +44,136 @@ static void CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outC
 
 	if (X != NULL) {
 		fftwf_free(X);
-		X = NULL;
 	}
 	if (Y != NULL) {
 		fftwf_free(Y);
-		Y = NULL;
+	}
+	if (in1 != NULL) {
+		fftwf_free(in1);
+	}
+	if (in2 != NULL) {
+		fftwf_free(in2);
+	}
+	if (chirpData != NULL) {
+		fftwf_free(chirpData);
 	}
 }
 
-static void reorderData(fftwf_complex *x, fftwf_complex *y, size_t bins, size_t channels) {
-	fftwf_complex *data[] = { x, y };
-	fftwf_complex tmp;
+static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac) {
+	const float dmFactConst = 2.0 * M_PI * coherentDM / dmPhaseConst;
+	const float chanbw = subbandbw / chanFac;
+	for (int32_t subband = 0; subband < nsub; subband++) {
+		const float subbandFreq = fbottom + subband * subbandbw + 0.5 * subbandbw;
+		for (int32_t chan = 0; chan < chanFac; chan++) {
+			const float channelFreq = subbandFreq - 0.5 * subbandbw + subbandbw * ((float) chan / (float) chanFac) + 0.5 * chanbw;
+			for (int32_t bin = 0; bin < mbin; bin++) {
+				const int32_t fftSpaceIdx = subband * mbin * chanFac + chan * mbin + bin;
 
-	size_t inputIdx, outputIdx;
+				// Frequency in FFT space
+				const float binFreq = chanbw * ((float) bin / (float) mbin) + 0.5 * (chanbw / mbin - chanbw);
 
-	for (int i = 0; i < 2; i++) {
-		fftwf_complex *workingPtr = data[i];
+				// Attempting to maintain similarity to CDMT, which is based upon
+				// Eq. 1 https://articles.adsabs.harvard.edu/pdf/2006ChJAS...6b..53L
+				const float phase = -1.0f * binFreq * binFreq * dmPhaseConst / ((channelFreq + binFreq) + channelFreq * channelFreq);
[…]
