### 2023-06-02 #5517 @editorialbot
[REVIEW]: udpPacketManager: An International LOFAR Station Data (Pre-)Processor
**Submitting author:** @David-McKenna (<a href="http://orcid.org/0000-0001-7185-1310">David McKenna</a>)
**Repository:** https://github.com/David-McKenna/udpPacketManager/
**Branch with paper.md** (empty if default branch):
**Version:** 0.9.2
