+++ src/CLI/lofar_cli_stokes.c
@@ -20,16 +20,31 @@ typedef enum stokes_t {
 	STOKESV = 8
 } stokes_t;
 
+typedef enum window_t {
+	// Eq. 1 https://articles.adsabs.harvard.edu/pdf/2006ChJAS...6b..53L
+	NO_WINDOW = 0,
+	COHERENT_DEDISP = 1,
+	PSR_STANDARD = 2,
+	BOXCAR = 4,
+
+} window_t;
+
 void helpMessages() {
 	printf("LOFAR Stokes extractor (CLI v%s, lib v%s)\n\n", UPM_CLI_VERSION, UPM_VERSION);
 	printf("Usage: lofar_cli_stokes <flags>");
 
+	helpMessages();
 	printf("\n\n");
 
 
-	printf("-C: <factor>    Channelisation factor to apply when processing data (default: disabled == 1)\n");
+	printf("-C <factor>     Channelisation factor to apply when processing data (default: disabled == 1)\n");
 	printf("-d <factor>     Temporal downsampling to apply when processing data (default: disabled == 1)\n");
-	printf("-D              Apply temporal downsampling to spectral data (slower, but higher quality (default: disabled)\n");
+	printf("-B <binFactor>  The count of FFT bins as a factor of the channelisation factor (default: 8, minimum: 3)\n");
+	printf("-N <factor>     The number of FFTs per channel to perform per iteration (default: 512)\n");
+	printf("-F              Apply temporal downsampling to spectral data (slower, but higher quality (default: disabled)\n");
+	printf("-D <dm>         Apply coherent dedispersion to the data (default: false; 0 pc/cc)\n");
+	printf("-W              Window function to apply to the data (defulat: PSR_STANDARD)\n");
+
 
 }
 
@@ -339,7 +354,7 @@ int main(int argc, char *argv[]) {
 	int8_t flagged = 0;
 
 	// FFTW strategy
-	int32_t channelisation = 1, downsampling = 1, spectralDownsample = 0;
+	int32_t channelisation = 1, downsampling = 1, spectralDownsample = 0, nfactor = 8, nforward = 512;
 	fftwf_complex *intermediateX = NULL;
 	fftwf_complex *intermediateY = NULL;
 	fftwf_complex * in1 = NULL;
@@ -352,7 +367,7 @@ int main(int argc, char *argv[]) {
 	int8_t stokesParameters = 0, numStokes = 0;
 
 	// Standard ugly input flags parser
-	while ((inputOpt = getopt(argc, argv, "rzqfvVDhi:o:m:M:I:u:t:s:S:b:C:d:P:T:")) != -1) {
+	while ((inputOpt = getopt(argc, argv, "crzqfvVFhD:i:o:m:M:I:u:t:s:S:b:C:d:P:T:B:N:")) != -1) {
 		input = 1;
 		switch (inputOpt) {
 
@@ -374,11 +389,6 @@ int main(int argc, char *argv[]) {
 				}
 				if (config->metadata_config.metadataType == NO_META) config->metadata_config.metadataType = lofar_udp_metadata_parse_type_output(optarg);
 				outputProvided = 1;
-				break;
-
-			case 'm':
-				config->packetsPerIteration = strtol(optarg, &endPtr, 10);
-				if (checkOpt(inputOpt, optarg, endPtr)) { flagged = 1; }
 				break;
 
 			case 'M':
@@ -428,6 +438,10 @@ int main(int argc, char *argv[]) {
 				config->replayDroppedPackets = 1;
 				break;
 
+			case 'c':
+				config->calibrateData = GENERATE_JONES;
+				break;
+
 			case 'C':
[…]
