+++ src/CLI/lofar_cli_stokes.c
@@ -74,8 +74,38 @@ CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_
 	}
 }
 
-static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac) {
-	const float dmFactConst = 2.0 * M_PI * coherentDM / dmPhaseConst;
+#pragma omp declare simd
+static inline void calibrateDataFunc(float *X, float *Y, const float *beamletJones, const int8_t **inputs,
+                                     const int64_t baseOffset) {
+	*(X) = calibrateSample(beamletJones[0], inputs[0][baseOffset], \
+                            -1.0f * beamletJones[1], inputs[0][baseOffset + 1], \
+                            beamletJones[2], inputs[1][baseOffset], \
+                            -1.0f * beamletJones[3], inputs[1][baseOffset + 1]);
+
+	*(&(X[1])) = calibrateSample(beamletJones[0], inputs[0][baseOffset + 1], \
+                            beamletJones[1], inputs[0][baseOffset], \
+                            beamletJones[2], inputs[1][baseOffset + 1], \
+                            beamletJones[3], inputs[1][baseOffset]);
+
+	*(Y) = calibrateSample(beamletJones[4], inputs[0][baseOffset], \
+                            -1.0f * beamletJones[5], inputs[0][baseOffset + 1], \
+                            beamletJones[6], inputs[1][baseOffset], \
+                            -1.0f * beamletJones[7], inputs[1][baseOffset + 1]);
+
+	*(&(X[1])) = calibrateSample(beamletJones[4], inputs[0][baseOffset + 1], \
+                            beamletJones[5], inputs[0][baseOffset], \
+                            beamletJones[6], inputs[1][baseOffset + 1], \
+                            beamletJones[7], inputs[1][baseOffset]);
+}
+
+static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac, window_t window) {
+	float dmFactConst;
+	if (window & COHERENT_DEDISP) {
+		dmFactConst = 2.0 * M_PI * coherentDM / dmPhaseConst;
+		window ^= 1; // Remove contribution
+	} else {
+		dmFactConst = 0;
+	}
 	const float chanbw = subbandbw / chanFac;
 	for (int32_t subband = 0; subband < nsub; subband++) {
 		const float subbandFreq = fbottom + subband * subbandbw + 0.5 * subbandbw;
@@ -86,14 +116,31 @@ static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, fl
 
 				// Frequency in FFT space
 				const float binFreq = chanbw * ((float) bin / (float) mbin) + 0.5 * (chanbw / mbin - chanbw);
+				float taperScale;
 
-				// Attempting to maintain similarity to CDMT, which is based upon
-				// Eq. 1 https://articles.adsabs.harvard.edu/pdf/2006ChJAS...6b..53L
-				const float phase = -1.0f * binFreq * binFreq * dmPhaseConst / ((channelFreq + binFreq) + channelFreq * channelFreq);
[…]
