### 2023-06-11 d782258
Check packet headers for malformed values (can typically happen during or shortly after a swlevel 2->3 transition if the station was interrupted)
