### 2023-06-11 cece94c
Cleaner implementation, TODO: reader side re-alignment, is the time major mode actually working?
