### 2023-06-26 c3cb99f
General documentation pass (following #12)
MkDocs Config (will be hosted at upm-docs.pulsar-observer)
Additional note on intended audience in main README
Fix-up some typos, missing JOSS information
Fix a mistake in README_CLI.md
