+++ docs/NEW_PROCESSING_MODE.md
@@ -2,33 +2,33 @@ Adding New Processing Modes
 ---------
 
 1. Add a named enum for the processing mode as a `processMode_t` in
-   `[lofar_udp_general.h]](../src/lib/lofar_udp_general.h.in)`, and to 
+   `[lofar_udp_general.h]](../src/lib/lofar_udp_general.h.in)`, and to
    the list of testables modes in
    `[lib_reference_files.hpp](../tests/lib_tests/lib_reference_files.hpp)`
    in order to makeit pass through the standard battery of tests.
 
 2. Create the CPP/C bridge `if--else` statements in
-   `[lofar_udp_backends.cpp](../src/lib/lofar_udp_backends.cpp)`, 
-   the main function is called `int32_t 
-   lofar_udp_cpp_loop_interface(lofar_udp_obs_meta *meta)`. You will 
-   need to pick both a processing mode int enum (any value greater 
-   than 0 and not in use by other modes) and an output data format. 
-   -- You will need to add the statement 6 times in total: with / 
-   without calibration (of disable calibration as an option) and for 
-   the 3 input bit modes, 4, 8 and 16. -- Calibration takes a 1 when 
-   enabled, 0 when disabled. -- Input type is always signed char for 
-   4-bit and 8-bit inputs, 16-bit takes signed short as the input. 
-   -- For copy methods, the output datatype should be the same as 
-   the input. Though you can change it, eg to convert to float by 
-   using float as the output datatype. Be sure to account for this 
-   later on when calculating output sizes. -- The 4-bit processing 
-   enum is (almost) always 4000 larger than the default enum, to 
-   signal to the processing loop that the data packet needs to have 
-   the bits upacked before proceeding. If you have a processing mode 
-   that just performs a type independant data move, e.g. 
+   `[lofar_udp_backends.cpp](../src/lib/lofar_udp_backends.cpp)`,
+   the main function is called `int32_t
+   lofar_udp_cpp_loop_interface(lofar_udp_obs_meta *meta)`. You will
+   need to pick both a processing mode int enum (any value greater
+   than 0 and not in use by other modes) and an output data format.
+   -- You will need to add the statement 6 times in total: with /
+   without calibration (of disable calibration as an option) and for
+   the 3 input bit modes, 4, 8 and 16. -- Calibration takes a 1 when
+   enabled, 0 when disabled. -- Input type is always signed char for
+   4-bit and 8-bit inputs, 16-bit takes signed short as the input.
+   -- For copy methods, the output datatype should be the same as
+   the input. Though you can change it, eg to convert to float by
+   using float as the output datatype. Be sure to account for this
+   later on when calculating output sizes. -- The 4-bit processing
+   enum is (almost) always 4000 larger than the default enum, to
+   signal to the processing loop that the data packet needs to have
+   the bits unpacked before proceeding. If you have a processing mode
+   that just performs a type independent data move, e.g.
    memcpy an entire packet, this change is not needed.
 
-  Here's an example of what mode 30, the time-majour single output 
+  Here's an example of what mode 30, the time-major single output
   looks like in the function.
 
 ```
@@ -83,13 +83,13 @@ switch (calibrateData) {
 }
 ```
 
-   If you miss a case, the default cases should raise an error when the 
-   tests are run, allowing you to indentify which case is missing 
-   and add it.
+  If you miss a case, the default cases should raise an error when the
+  tests are run, allowing you to identify which case is missing
+  and add it.
 
-3. Create the task kernel in `[lofar_udp_backends.hpp](../src/lib/lofar_udp_backends.hpp)`, following the 
[…]
