### 2023-06-26 d0418aa
Typo fix; further Docker/Singularity information
