### 2023-06-28 #5517 @pritchardn
A few minor comments on the Software Paper (mainly typos) which is otherwise excellent:

- [29] - pecular -> peculiar
- [43] - 2 times -> two times
- [61-66] - I suggest mentioning the names of the CLIs directly to make it completely clear which you refer to
