### 2023-06-28 #5517 @David-McKenna
Hey @pritchardn,

Thanks! I've made changes to the paper and the general documentation comments you made in the docsPass branch (PR David-McKenna/udpPacketManager#13 ).

As for the PSRDADA issue, are sourceforge's certs out of date again? That's been a chronic issue over the past few years. I'll see if there's another way for me to download their source in my CMake script.

Cheers,
David
