### 2023-06-28 b47c3f1
Incorporate @pritchardn's feedback from openjournals/joss-reviews#5517
