+++ paper/paper.md
@@ -62,7 +62,7 @@ data to a PUMA2-derived data format that can be parsed using the DSPSR
 [@dspsr] ecosystem, and the ARTEMIS system [@artemis], a 
 hardware-software package for online transient observations derived 
 from the source-available `PELICAN` LOFAR backend [@pelican]. These 
-were found not to be sufficiently flexible to account for percular 
+were found not to be sufficiently flexible to account for peculiar 
 ways of utilising the telescope hardware, such as supporting 
 multi-mode observations [@mckb357], nor work within some constraints 
 of the REALTA compute cluster [@realta]. Consequently, 
@@ -84,7 +84,7 @@ Postdam, in 2022 [@louisJupiter].
 multiple different data sources and sinks in a semi-transparent 
 manner, allowing for data to be read and written out through normal 
 files, FIFO named pipes, `zstandard` [@zstd] compressed files (which 
-allow for a greater than 2 times space-saving when handling raw 
+allow for a greater than two times space-saving when handling raw 
 voltages from LOFAR stations), `PSRDADA` ringbuffers for online 
 processing [@psrdada], and write-only to `HDF5` files that conform to 
 LOFAR ICD003 [@icd003]. These interfaces were also used to create 
@@ -112,13 +112,14 @@ by applying, or generating for external use, Jones matrices generated by
 `dreamBeam` [@dreamBeam].
 
 Two command-line-interfaces are provided alongside the library. The 
-first offers an easy interface for accessing the normal 
-functionality of the library for observers, while the second utilises 
-the output voltages to perform additional channelisation of Stokes 
-parameters through the `FFTW` library [@fftw] and additional temporal 
-downsampling beyond the normal factor of 16 limit. Both of these act 
-further act as examples of how to further utilise the library and 
-its outputs.
+first, `lofar_udp_extractor`, offers an easy interface for accessing 
+the normal functionality of the library for observers, while the second,
+`lofar_stokes_extractor`, utilises the output voltages to perform 
+additional processing for science-ready outputs, such as channelisation 
+of Stokes parameters through the `FFTW` library [@fftw] and additional 
+temporal downsampling beyond the library's factor of 16 limit. Both of 
+these interfaces act further act as examples of how to further utilise 
+the library and its outputs.
 
 # Acknowledgements
 
