### 2023-06-28 47fe23b
Move pip install to a requirements.txt file for easy out-of-build calls
Allows for `pip3 install -r requirements.txt` when setting up a new venv
for processing runs, without digging into CMakeLists.txt or manually
installing dependencies mentioned in the README
