### 2023-06-28 7a37fdf
Change to git protocol for Sourceforge repo (PSRDADA) until a better solution is found
Works around the chronic CAfile issue, unfortunately the git protocol is
significantly less secure than HTTPS/ssh, but it's the only other option
for cloning data from Sourceforge. Their snapshot downloads are not
guaranteed to persist beyond the commit existing as the HEAD of a
branch, and do not contain the metadata needed for a `git reset`.
