### 2023-06-29 03460f7
Minor calibration fixes
Bitmode was incorrectly swapped in some cases
Better checks on calibration capability based on the enum, not values
