+++ src/lib/lofar_udp_structs_metadata.c
@@ -75,7 +75,7 @@ const lofar_udp_metadata lofar_udp_metadata_default = {
 	.upm_procmode = UNSET_MODE,
 	.upm_bandflip = -1,
 	.upm_replay = -1,
-	.upm_calibrated = -1,
+	.upm_calibrated = NO_CALIBRATION,
 	.upm_blocksize = -1,
 	.upm_pack_per_iter = -1,
 	.upm_processed_packets = -1,
