### 2023-06-29 69d1247
Add additional bitmode checks to the reader mode tests, silence some debug outputs
DADA header hashes change as a result of changing the instrument key from CASPSR to LOFAR
