### 2023-06-29 05bba80
Documentation pass (#13)
* General documentation pass (following #12)

MkDocs Config (will be hosted at upm-docs.pulsar-observer)
Additional note on intended audience in main README
Fix-up some typos, missing JOSS information
Fix a mistake in README_CLI.md

* Typo fix; further Docker/Singularity information

* Incorporate @pritchardn's feedback from openjournals/joss-reviews#5517

* Change outdated help message for "-c" in CLIs
