+++ docs/README_io.md
@@ -1,66 +1,66 @@
 # Overview
 
-The `lofar_udp_io_*` function offer an interface to one of several 
-I/O sources, from writing to disk as normal files or Zstandard 
-compressed files, or to in-memory writes through named FIFO pipes 
-and PSRDADA ringbuffers. 
+The `lofar_udp_io_*` function offer an interface to one of several
+I/O sources, from writing to disk as normal files or Zstandard
+compressed files, or to in-memory writes through named FIFO pipes
+and PSRDADA ringbuffers.
 
-The standard way to access the interfaces are through the use of a 
-`_alloc()` and then a `_setup()` function to configure a struct for 
-I/O operations, followed by several calls to a read/write 
-operation function, and then a `_cleanup()` function before exiting 
-the program. The sections below give a brief overview of the 
-expected configuration of the structs before calling the setup 
-functions, and the expected behaviour while performing some I/O 
+The standard way to access the interfaces are through the use of a
+`_alloc()` and then a `_setup()` function to configure a struct for
+I/O operations, followed by several calls to a read/write
+operation function, and then a `_cleanup()` function before exiting
+the program. The sections below give a brief overview of the
+expected configuration of the structs before calling the setup
+functions, and the expected behaviour while performing some I/O
 operations.
 
 # `lofar_udp_io_read` Interface
 
-## Setup 
-A pre-initialised struct can be allocated on the heap by calling the 
-`lofar_udp_io_read_config_alloc()` function. It may return `NULL` if 
+## Setup
+A pre-initialised struct can be allocated on the heap by calling the
+`lofar_udp_io_read_config_alloc()` function. It may return `NULL` if
 the allocation fails.
 
-The struct should then be configured, with at least the following 
+The struct should then be configured, with at least the following
 members set:
 - readerType
 - \(inputLocations OR inputDadaKeys\)\[0:numInputs\]
 - numInputs
 
-There is a function for parsing the first and second variables directly from an input `optarg` flag or char array into the 
-configuration struct, `lofar_udp_io_read_parse_optarg()`. This expects an input string following the conventions that are detailed in the [the 
+There is a function for parsing the first and second variables directly from an input `optarg` flag or char array into the
+configuration struct, `lofar_udp_io_read_parse_optarg()`. This expects an input string following the conventions that are detailed in the [the
 CLI readme](README_CLI.md).
 
 Additionally, we recommend preparing:
 - The maximum number of elements expected to be read per iteration
 - A buffer than can hold this number of elements
 
-Both of those variables can then be used alongside the 
-`lofar_udp_io_read_setup_helper` function in order to simplify the 
-setup process, accounting for any requirements of some processing 
+Both of those variables can then be used alongside the
+`lofar_udp_io_read_setup_helper` function in order to simplify the
+setup process, accounting for any requirements of some processing
 methods (such as the page sizes for Zstandard compressed buffers).
 
-If you want to directly use the `lofar_udp_io_read_setup` function, 
+If you want to directly use the `lofar_udp_io_read_setup` function,
 the size variable should be set into the `readBufSize[0,
-numInputs]`, `decompressionTracker[0,numInputs].size`, while the 
-buffer is used to initialise the `decompressionTracker[port].src` if 
+numInputs]`, `decompressionTracker[0,numInputs].size`, while the
+buffer is used to initialise the `decompressionTracker[port].src` if
 the readerType is `ZSTDCOMPRESSED`.
 
 ## Operations
 
-Repeated calls to `lofar_udp_io_read`, alongside a target buffer and requested read length will produce as much data as possible from the 
+Repeated calls to `lofar_udp_io_read`, alongside a target buffer and requested read length will produce as much data as possible from the
[…]
