### 2023-06-29 a572a52
DADA & Calibration Header Fixes (#16)
* Only run tests on oldest/newest compiler verisons to speed up CI

* Minor calibration fixes

Bitmode was incorrectly swapped in some cases
Better checks on calibration capability based on the enum, not values

* Change default instrument from CASPSR to LOFAR

* Add documentation on generating generic DADA voltage output (for DSPSR)

* Spellcheck on new docs

* Add additional bitmode checks to the reader mode tests, silence some debug outputs

DADA header hashes change as a result of changing the instrument key from CASPSR to LOFAR
