### 2023-06-29 07a839f
Improve build stability (#14) (git:// security concern)
* Move pip install to a requirements.txt file for easy out-of-build calls

Allows for `pip3 install -r requirements.txt` when setting up a new venv
for processing runs, without digging into CMakeLists.txt or manually
installing dependencies mentioned in the README

* Add prefixes where missing to source paths in CMakeLists.txt

* Change to git protocol for Sourceforge repo (PSRDADA) until a better solution is found

Works around the chronic CAfile issue, unfortunately the git protocol is
significantly less secure than HTTPS/ssh, but it's 
[…]
