### 2023-06-30 cd4b77b
Add an option for single-output (auto-)correlations
