+++ src/CLI/lofar_cli_stokes.c
@@ -17,7 +17,8 @@ typedef enum stokes_t {
 	STOKESI = 1,
 	STOKESQ = 2,
 	STOKESU = 4,
-	STOKESV = 8
+	STOKESV = 8,
+	CORRLTE = 16,
 } stokes_t;
 
 typedef enum window_t {
@@ -75,6 +76,14 @@ CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_
 }
 
 #pragma omp declare simd
+static inline void generateVoltageCorrelations(fftwf_complex X, fftwf_complex Y, float *accumulations) {
+	accumulations[0] = (X[0] * X[0]) + (X[1] * X[1]);
+	accumulations[1] = (Y[0] * Y[0]) + (Y[1] * Y[1]);
+	accumulations[2] = ((X[0] * Y[0]) + (X[1] * Y[1]));
+	accumulations[3] = ((X[0] * Y[1]) - (X[1] * Y[0]));
+}
+
+#pragma omp declare simd
 static inline void calibrateDataFunc(float *X, float *Y, const float *beamletJones, const int8_t **inputs,
                                      const int64_t baseOffset) {
 	*(X) = calibrateSample(beamletJones[0], inputs[0][baseOffset], \
@@ -98,7 +107,7 @@ static inline void calibrateDataFunc(float *X, float *Y, const float *beamletJon
                             beamletJones[7], inputs[1][baseOffset]);
 }
 
-static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac, window_t window) {
+static int8_t windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac, window_t window) {
 	float dmFactConst;
 	if (window & COHERENT_DEDISP) {
 		dmFactConst = 2.0 * M_PI * coherentDM / dmPhaseConst;
@@ -129,11 +138,13 @@ static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, fl
 						break;
 
 					default:
+						fprintf(stderr, "ERROR: No window selected for %s (window %d), exiting.\n", __func__, window);
+						return 1;
 						break;
 				}
 
 
-				if (dmFactConst) {
+				if (window & COHERENT_DEDISP) {
 					const float phase = -1.0f * binFreq * binFreq * dmPhaseConst / ((channelFreq + binFreq) + channelFreq * channelFreq);
 					chirpFunc[fftSpaceIdx][0] = cos(phase) * taperScale;
 					chirpFunc[fftSpaceIdx][1] = cos(phase) * taperScale;
@@ -144,6 +155,8 @@ static void windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, fl
 			}
 		}
 	}
+
+	return 0;
 }
 
 #define npol 2
@@ -280,6 +293,7 @@ transposeDetect(fftwf_complex (*X), fftwf_complex (*Y), float **outputs, size_t 
 	noverlap /= channelisation;
 	const size_t outputNchan = nsub * channelisation / channelDownsample;
 	const size_t outputMbin = mbin - 2 * noverlap;
+	const int64_t correlateScale = (stokesFlags & CORRLTE) ? UDPNPOL : 1;
 
 #pragma omp parallel for default(shared)
 	for (size_t sub = 0; sub < nsub; sub++) {
@@ -299,46 +313,61 @@ transposeDetect(fftwf_complex (*X), fftwf_complex (*Y), float **outputs, size_t 
 					// Input is time major
 					//         curr     fftoffset       total samples          size per channel
[…]
