### 2023-06-30 3bab0f9
Add additional window checks
