### 2023-07-03 cf5179b
Fix memory leak from outputStokes array and meta->headerBuffer, remove headerBuffer in Stokes CLI
