+++ src/CLI/lofar_cli_stokes.c
@@ -50,12 +50,11 @@ void helpMessages() {
 }
 
 static void
-CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_t *header, fftwf_complex (*X), fftwf_complex (*Y), fftwf_complex *in1,
-           fftwf_complex (*in2), fftwf_complex (*chirpData)) {
+CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, fftwf_complex (*X), fftwf_complex (*Y), fftwf_complex (*in1), fftwf_complex (*in2),
+           fftwf_complex (*chirpData), float **outputStokes) {
 
 	FREE_NOT_NULL(outConfig);
 	FREE_NOT_NULL(config);
-	FREE_NOT_NULL(header);
 
 
 	if (X != NULL) {
@@ -72,6 +71,12 @@ CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, int8_
 	}
 	if (chirpData != NULL) {
 		fftwf_free(chirpData);
+	}
+
+	if (outputStokes != NULL) {
+		for (int32_t i = 0; i < MAX_OUTPUT_DIMS; i++) {
+			FREE_NOT_NULL(outputStokes[i]);
+		}
 	}
 }
 
@@ -427,8 +432,6 @@ int main(int argc, char *argv[]) {
 	lofar_udp_config *config = lofar_udp_config_alloc();
 	lofar_udp_io_write_config *outConfig = lofar_udp_io_write_alloc();
 
-	int8_t *headerBuffer = NULL;
-
 	if (config == NULL || outConfig == NULL) {
 		fprintf(stderr, "ERROR: Failed to allocate memory for configuration structs (something has gone very wrong...), exiting.\n");
 		FREE_NOT_NULL(config); FREE_NOT_NULL(outConfig);
@@ -470,7 +473,7 @@ int main(int argc, char *argv[]) {
 			case 'i':
 				if (strncpy(inputFormat, optarg, DEF_STR_LEN - 1) != inputFormat) {
 					fprintf(stderr, "ERROR: Failed to store input data file format, exiting.\n");
-					CLICleanup(config, outConfig, headerBuffer, intermediateX, intermediateY, in1, in2, chirpData);
+					CLICleanup(config, outConfig, intermediateX, intermediateY, in1, in2, chirpData, NULL);
 					return 1;
 				}
 				inputProvided = 1;
@@ -480,7 +483,7 @@ int main(int argc, char *argv[]) {
 			case 'o':
 				if (lofar_udp_io_write_parse_optarg(outConfig, optarg) < 0) {
 					helpMessages();
-					CLICleanup(config, outConfig, headerBuffer, intermediateX, intermediateY, in1, in2, chirpData);
+					CLICleanup(config, outConfig, intermediateX, intermediateY, in1, in2, chirpData, NULL);
 					return 1;
 				}
 				if (config->metadata_config.metadataType == NO_META) config->metadata_config.metadataType = lofar_udp_metadata_parse_type_output(optarg);
@@ -494,7 +497,7 @@ int main(int argc, char *argv[]) {
 			case 'I':
 				if (strncpy(config->metadata_config.metadataLocation, optarg, DEF_STR_LEN) != config->metadata_config.metadataLocation) {
 					fprintf(stderr, "ERROR: Failed to copy metadata file location to config, exiting.\n");
-					CLICleanup(config, outConfig, headerBuffer, intermediateX, intermediateY, in1, in2, chirpData);
+					CLICleanup(config, outConfig, intermediateX, intermediateY, in1, in2, chirpData, NULL);
 					return 1;
 				}
 				break;
@@ -507,7 +510,7 @@ int main(int argc, char *argv[]) {
[…]
