### 2023-07-03 f7626f5
Fix incorrect data offsets in overlapAndPad
