### 2023-07-03 8999665
Fix writeOffset to be in floats rather than bytes
