### 2023-07-03 7171dca
omp_get_num_threads -> omp_get_max_threads (how did backends.hpp never segfault on 4-bit data before?)
