### 2023-07-03 5cecdab
Better NULL-checks for outputStokes (prevents segfaults in Release builds)
