### 2023-07-03 1efd964
Add explicit null-checks to beamletJones in overlapAndPad and padNextIteration
For some reason, despite an explicit set to NULL, this was being treated as a true statement in Release mode.
