### 2023-07-03 b9e7836
The compiler isn't always right: Remove omp simd pragmas
For some reason, when present they cause the wrong branch to be taken for the inner beamletJones if statement
