### 2023-07-03 e81c0df
Missed a pragma omp simd in the last commit
