### 2023-07-03 819aa6f
Move debug messages to verbose mode
