### 2023-07-03 0f2ce0c
Correctly apply coherent dedispersion when requested
