### 2023-07-03 a9656df
Fix time-based downsampling (was so focused on spectral downsampling I completely missed this)
