### 2023-07-04 9b956dc
Fix nfft/nforward issues, general code cleanup
