+++ src/CLI/lofar_cli_stokes.c
@@ -82,10 +82,10 @@ CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, fftwf
 
 #pragma omp declare simd
 static inline void generateVoltageCorrelations(fftwf_complex X, fftwf_complex Y, float *accumulations) {
-	accumulations[0] = (X[0] * X[0]) + (X[1] * X[1]);
-	accumulations[1] = (Y[0] * Y[0]) + (Y[1] * Y[1]);
-	accumulations[2] = ((X[0] * Y[0]) + (X[1] * Y[1]));
-	accumulations[3] = ((X[0] * Y[1]) - (X[1] * Y[0]));
+	accumulations[0] += (X[0] * X[0]) + (X[1] * X[1]);
+	accumulations[1] += (Y[0] * Y[0]) + (Y[1] * Y[1]);
+	accumulations[2] += ((X[0] * Y[0]) + (X[1] * Y[1]));
+	accumulations[3] += ((X[0] * Y[1]) - (X[1] * Y[0]));
 }
 
 #pragma omp declare simd
@@ -112,33 +112,36 @@ static inline void calibrateDataFunc(float *X, float *Y, const float *beamletJon
                             beamletJones[7], inputs[1][baseOffset]);
 }
 
-static int8_t windowGenerator(fftwf_complex *const chirpFunc, float coherentDM, float fbottom, float subbandbw, int32_t mbin, int32_t nsub, int32_t chanFac, window_t window) {
-	float dmFactConst;
+static int8_t windowGenerator(fftwf_complex *const chirpFunc, const double coherentDM, const double fbottom, const double subbandbw, const int32_t mbin, const int32_t nsub, const int32_t chanFac, window_t window) {
+	double dmFactConst;
 	int8_t coherentDedisp;
 	if (window & COHERENT_DEDISP) {
 		dmFactConst = 2.0 * M_PI * coherentDM / dmPhaseConst;
-		window ^= 1; // Remove contribution
+		window ^= COHERENT_DEDISP; // Remove contribution
 		coherentDedisp = 1; // Set flag (float comparisons are unstable)
 	} else {
 		dmFactConst = 0;
 		coherentDedisp = 0;
 	}
-	const float chanbw = subbandbw / chanFac;
+	const double chanbw = subbandbw / (double) chanFac;
+
 	for (int32_t subband = 0; subband < nsub; subband++) {
-		const float subbandFreq = fbottom + subband * subbandbw + 0.5 * subbandbw;
+		const double subbandFreq = fbottom + (double) subband * subbandbw + 0.5 * subbandbw;
+
 		for (int32_t chan = 0; chan < chanFac; chan++) {
-			const float channelFreq = subbandFreq - 0.5 * subbandbw + subbandbw * ((float) chan / (float) chanFac) + 0.5 * chanbw;
+			const double channelFreq = subbandFreq - 0.5 * subbandbw + subbandbw * ((double) chan / (double) chanFac) + 0.5 * chanbw;
+
 			for (int32_t bin = 0; bin < mbin; bin++) {
 				const int32_t fftSpaceIdx = subband * mbin * chanFac + chan * mbin + bin;
 
 				// Frequency in FFT space
-				const float binFreq = chanbw * ((float) bin / (float) mbin) + 0.5 * (chanbw / mbin - chanbw);
-				float taperScale;
+				const double binFreq = chanbw * ((double) bin / (double) mbin) + 0.5 * (chanbw / (double) mbin - chanbw);
+				double taperScale;
 
 
 				switch (window) {
 					case PSR_STANDARD:
[…]
