### 2023-07-05 bfa9325
_lofar_udp_metadata_setup_types: Allow for metadata refreshing after manual overrides
