### 2023-07-05 ea8f45c
Temporary safety measure: Just exit on partial/end block reads.
