### 2023-07-05 e7d62b9
ZSTD Reader: Tests for optimising madvice calls; bring back free/dontneed call on read
