+++ src/lib/io/lofar_udp_io_ZSTD.c
@@ -46,7 +46,7 @@ _lofar_udp_io_read_setup_ZSTD(lofar_udp_io_read_config *const input, const char 
 		return -2;
 	}
 
-	if (madvise(tmpPtr, fileSize, MADV_SEQUENTIAL) == -1) {
+	if (madvise(tmpPtr, fileSize, MADV_SEQUENTIAL | MADV_WILLNEED) == -1) {
 		fprintf(stderr, "ERROR: Failed to advise the kernel on mmap read strategy on port %d. Errno: %d. Exiting.\n",
 				port, errno);
 		return -3;
@@ -56,14 +56,14 @@ _lofar_udp_io_read_setup_ZSTD(lofar_udp_io_read_config *const input, const char 
 	// Setup the decompressed data buffer/struct
 	// If the size is pre-set, we have already resize the true buffer as required
 	int64_t bufferSize = input->decompressionTracker[port].size != 0 ?
-	                     (int64_t) input->decompressionTracker[port].size : input->readBufSize[port];
+						 (int64_t) input->decompressionTracker[port].size : input->readBufSize[port];
 	if (bufferSize % ZSTD_DStreamOutSize() && !(input->readerType == ZSTDCOMPRESSED_INDIRECT)) {
 		fprintf(stderr, "ERROR %s: Zstandard buffer was not initialised correctly, exiting.\n", __func__);
 		return -1;
 	} else {
 		bufferSize += _lofar_udp_io_read_ZSTD_fix_buffer_size(bufferSize, 1);
 		VERBOSE(printf("reader_setup: expanding decompression buffer by %ld bytes\n",
-		               bufferSize - input->readBufSize[port]));
+					   bufferSize - input->readBufSize[port]));
 	}
 
 	input->decompressionTracker[port].pos = 0; // Initialisation for our step-by-step reader
@@ -110,7 +110,7 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 	VERBOSE(printf("reader_nchars %d: start of ZSTD read loop, %ld, %ld, %ld, %ld, %ld\n",
 				   port,
 				   input->readingTracker[port].pos,
-	               input->readingTracker[port].size,
+				   input->readingTracker[port].size,
 				   input->decompressionTracker[port].pos,
 				   dataRead,
 				   nchars););
@@ -149,10 +149,10 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 		previousDecompressionPos = input->decompressionTracker[port].pos;
 		// zstd streaming decompression + check for errors
 		returnVal = ZSTD_decompressStream(input->dstream[port], &(input->decompressionTracker[port]),
-		                                  &(input->readingTracker[port]));
+										  &(input->readingTracker[port]));
 		if (ZSTD_isError(returnVal)) {
 			fprintf(stderr, "ZSTD encountered an error decompressing a frame (code %ld, %s), exiting data read early.\n",
-			        returnVal, ZSTD_getErrorName(returnVal));
+					returnVal, ZSTD_getErrorName(returnVal));
 			break;
 		}
 
@@ -168,8 +168,8 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 
 		if (input->decompressionTracker[port].pos == input->decompressionTracker[port].size) {
 			fprintf(stderr,
-			        "Failed to read %ld/%ld chars on port %d before filling the buffer. Attempting to continue...\n",
-			        dataRead, nchars, port);
[…]
