### 2023-07-06 93c26a1
Move FFTW variables to a heap allocated struct for easier cleanup
