+++ src/CLI/lofar_cli_stokes.c
@@ -30,7 +30,19 @@ typedef enum window_t {
 
 } window_t;
 
-void helpMessages() {
+struct fftwVars {
+	fftwf_complex *intermediateX;
+	fftwf_complex *intermediateY;
+	fftwf_complex *in1;
+	fftwf_complex *in2;
+	fftwf_complex *chirpData;
+	fftwf_plan fftForwardX;
+	fftwf_plan fftForwardY;
+	fftwf_plan fftBackwardX;
+	fftwf_plan fftBackwardY;
+};
+
+void helpMessages(void) {
 	printf("LOFAR Stokes extractor (CLI v%s, lib v%s)\n\n", UPM_CLI_VERSION, UPM_VERSION);
 	printf("Usage: lofar_cli_stokes <flags>");
 
@@ -50,27 +62,28 @@ void helpMessages() {
 }
 
 static void
-CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, fftwf_complex (*X), fftwf_complex (*Y), fftwf_complex (*in1), fftwf_complex (*in2),
-           fftwf_complex (*chirpData), float **outputStokes) {
+CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, float **outputStokes, struct fftwVars *fftw) {
 
 	FREE_NOT_NULL(outConfig);
 	FREE_NOT_NULL(config);
 
 
-	if (X != NULL) {
-		fftwf_free(X);
-	}
-	if (Y != NULL) {
-		fftwf_free(Y);
-	}
-	if (in1 != NULL) {
-		fftwf_free(in1);
-	}
-	if (in2 != NULL) {
-		fftwf_free(in2);
-	}
-	if (chirpData != NULL) {
-		fftwf_free(chirpData);
+	if (fftw != NULL) {
+		ARB_FREE_NOT_NULL(fftw->intermediateX, fftwf_free);
+		ARB_FREE_NOT_NULL(fftw->intermediateY, fftwf_free);
+		ARB_FREE_NOT_NULL(fftw->in1, fftwf_free);
+		ARB_FREE_NOT_NULL(fftw->in2, fftwf_free);
+		ARB_FREE_NOT_NULL(fftw->chirpData, fftwf_free);
+
+		ARB_FREE_NOT_NULL(fftw->fftForwardX, fftwf_destroy_plan);
+		ARB_FREE_NOT_NULL(fftw->fftForwardY, fftwf_destroy_plan);
+		ARB_FREE_NOT_NULL(fftw->fftBackwardX, fftwf_destroy_plan);
+		ARB_FREE_NOT_NULL(fftw->fftBackwardY, fftwf_destroy_plan);
+
+		fftwf_cleanup_threads();
+		fftwf_cleanup();
+
+		free(fftw);
 	}
 
 	if (outputStokes != NULL) {
@@ -461,17 +474,10 @@ int main(int argc, char *argv[]) {
 	char *endPtr;
 	int8_t flagged = 0;
 
+	struct fftwVars *fftw = calloc(sizeof(struct fftwVars), 1);
+
 	// FFTW strategy
 	int32_t channelisation = 1, downsampling = 1, spectralDownsample = 0, nfactor = 8, nforward = 512;
-	fftwf_complex *intermediateX = NULL;
-	fftwf_complex *intermediateY = NULL;
-	fftwf_complex * in1 = NULL;
-	fftwf_complex * in2 = NULL;
-	fftwf_complex * chirpData = NULL;
-	fftwf_plan fftForwardX;
-	fftwf_plan fftForwardY;
-	fftwf_plan fftBackwardX;
-	fftwf_plan fftBackwardY;
[…]
