### 2023-07-06 71dde0b
Reorder variables in cleanup func to better match usage
