+++ src/CLI/lofar_cli_stokes.c
@@ -62,7 +62,7 @@ void helpMessages(void) {
 }
 
 static void
-CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, float **outputStokes, struct fftwVars *fftw) {
+CLICleanup(lofar_udp_config *config, lofar_udp_io_write_config *outConfig, struct fftwVars *fftw, float **outputStokes) {
 
 	FREE_NOT_NULL(outConfig);
 	FREE_NOT_NULL(config);
@@ -488,7 +488,7 @@ int main(int argc, char *argv[]) {
 			case 'i':
 				if (strncpy(inputFormat, optarg, DEF_STR_LEN - 1) != inputFormat) {
 					fprintf(stderr, "ERROR: Failed to store input data file format, exiting.\n");
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				inputProvided = 1;
@@ -498,7 +498,7 @@ int main(int argc, char *argv[]) {
 			case 'o':
 				if (lofar_udp_io_write_parse_optarg(outConfig, optarg) < 0) {
 					helpMessages();
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				if (config->metadata_config.metadataType == NO_META) config->metadata_config.metadataType = lofar_udp_metadata_parse_type_output(optarg);
@@ -512,7 +512,7 @@ int main(int argc, char *argv[]) {
 			case 'I':
 				if (strncpy(config->metadata_config.metadataLocation, optarg, DEF_STR_LEN) != config->metadata_config.metadataLocation) {
 					fprintf(stderr, "ERROR: Failed to copy metadata file location to config, exiting.\n");
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				break;
@@ -525,7 +525,7 @@ int main(int argc, char *argv[]) {
 			case 't':
 				if (strncpy(inputTime, optarg, 255) != inputTime) {
 					fprintf(stderr, "ERROR: Failed to copy start time from input, exiting.\n");
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				break;
@@ -543,7 +543,7 @@ int main(int argc, char *argv[]) {
 			case 'b':
 				if (sscanf(optarg, "%hd,%hd", &(config->beamletLimits[0]), &(config->beamletLimits[1])) < 0) {
 					fprintf(stderr, "ERROR: Failed to scan input beamlets, exiting.\n");
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				break;
@@ -592,7 +592,7 @@ int main(int argc, char *argv[]) {
 				if (checkOpt(inputOpt, optarg, endPtr)) { flagged = 1; }
 				if (!flagged && nfactor < 6) {
 					fprintf(stderr, "ERROR: nfactor must be at least 6 due to FFT overlaps. Exiting.\n");
-					CLICleanup(config, outConfig, NULL, fftw);
+					CLICleanup(config, outConfig, fftw, NULL);
 					return 1;
 				}
 				break;
@@ -615,7 +615,7 @@ int main(int argc, char *argv[]) {
 			case 'P':
 				if (numStokes > 0) {
[…]
