### 2023-07-06 5b428fc
Error check FFTW plan/Stokes output array generation
