### 2023-07-06 9e14070
Always run FFTs if coherent dedispersion is requested
