### 2023-07-06 0610d68
Optimise temporal downsampling
