### 2023-07-06 fb1985a
Bump to C++20 for VA_OPT support in clang (GCC already had it), run tests on supported compiler matrix
