### 2023-07-06 bb27224
Revert last commit, use ##__VA_ARGS__ instead of __VA_OPT__ (seems to work on a few tests clang/gcc versions at least)
