### 2023-07-06 d2505e7
FFTW plan NULL checks was missing a comparison
