### 2023-07-06 ae9866b
Fix bandpass shift when using temporal downsampling
