### 2023-07-06 a7a1c0d
ZSTD read madvise optimisations: Only mark the recently read bytes
