+++ src/lib/io/lofar_udp_io_ZSTD.c
@@ -104,8 +104,11 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 	size_t returnVal;
 
 	// Move the trailing end of the buffer to the start of the target read
-	int8_t *dest = input->decompressionTracker[port].dst;
+	const int8_t *dest = input->decompressionTracker[port].dst;
 	dataRead = input->decompressionTracker[port].pos - input->zstdLastRead[port];
+
+	// Copy the preread position to optimise madvise calls later
+	const int64_t previousReadPos = input->readingTracker[port].pos;
 
 	VERBOSE(printf("reader_nchars %d: start of ZSTD read loop, %ld, %ld, %ld, %ld, %ld\n",
 				   port,
@@ -117,7 +120,7 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 
 	if (input->readerType == ZSTDCOMPRESSED) {
 		dest = targetArray;
-		int64_t ptrByteDifference = targetArray - (int8_t *) input->decompressionTracker[port].dst;
+		const int64_t ptrByteDifference = targetArray - (int8_t *) input->decompressionTracker[port].dst;
 		if (ptrByteDifference < 0 || ptrByteDifference > (int64_t) input->decompressionTracker[port].size) {
 			fprintf(stderr, "ERROR %s: Passed buffer is not within pre-set buffer range (delta %ld), exiting.\n", __func__, ptrByteDifference);
 			return -1;
@@ -187,7 +190,7 @@ int64_t _lofar_udp_io_read_ZSTD(lofar_udp_io_read_config *const input, int8_t po
 	// Previously commented out at it wasn't thought the be needed, unfortunately it is.
 	// Large observations were holding > 50% of a systems memory, as compared to < 3GB in 0.6 builds.
 	//if (madvise(((void *) input->readingTracker[port].src), input->readingTracker[port].pos, MADV_DONTNEED) < 0) {
-	if (madvise(((void *) input->readingTracker[port].src), input->readingTracker[port].pos, input->zstdMadvise) < 0) {
+	if (madvise(((void *) input->readingTracker[port].src) + previousReadPos, input->readingTracker[port].pos, input->zstdMadvise) < 0) {
 		fprintf(stderr,
 				"ERROR: Failed to apply %d after read operation on port %d (errno %d: %s).\n", input->zstdMadvise, port,
 				errno, strerror(errno));
