### 2024-01-26 #5517 @plaplant
@dfm thanks for checking in!

@David-McKenna I recently built the package for my review, and I started by trying to build it inside of an Ubuntu image on my ARM-based Mac using [multipass](https://multipass.run/). The compilation failed, but I think it was due to not being an x86-based machine. Building on an x86 system worked as intended, but I did not have enough privileges on that machine to run the tests. That's the reason I haven't checked the "functionality verification" box in my review. (Also, I'm not sure if it was a problem in my setup, or something explicitly about needing an x86 system. If the latter, maybe that's worth mentioning briefly in the docs?)

I saw as part of the GitHub actions the tests are being run there, so I'd be happy to count that as verifying the functionality. If something more independent is required, I'll have to poke around and see if I can get access to a suitable machine. (This is perhaps indirectly a question for @dfm.)
