### 2024-01-29 #5517 @dfm
> I saw as part of the GitHub actions the tests are being run there, so I'd be happy to count that as verifying the functionality. If something more independent is required, I'll have to poke around and see if I can get access to a suitable machine.

Thanks for bringing this up, @plaplant. From my perspective, I think this is okay as long as the architecture constraints are clearly documented!
