### 2024-02-01 #5517 @pritchardn
Hi all, I've tried again (successfully) to install locally and run the included tests.
I'm happy to pass this one :+1:
