### 2024-02-02 #5517 @David-McKenna
Hi all,

Apologies for my silence over the past few days.

I'm currently in the middle of moving country and starting a new role, so I
won't have a chance to address the above requests for another few days,
though I hope to get to them by the end of the week.

Cheers,
David

On Thu, 1 Feb 2024 at 04:10, Nicholas Pritchard ***@***.***>
wrote:

> Hi all, I've tried again (successfully) to install locally and run the
> included tests.
> I'm happy to pass this one 👍
>
