### 2024-02-26 58a2375
Add a note on supported architectures (following discussion in openjournals/joss-reviews#5517)
