### 2024-02-26 #5517 @David-McKenna
Apologies for the delay, I was expecting to get a work laptop with an ARM CPU in the past weeks so I could test the described issues, but there's been a hold-up with delivery. For the time being, I've added an additional note that currently only x86-64 architectures are supported ([58a2375](https://github.com/David-McKenna/udpPacketManager/commit/58a23752c54f6eadb68e62b51f9502d2a9f9b851)).

Cheers,
David
