### 2024-04-22 0e84cb8
Fixing some small typos (#18)
