### 2024-04-22 #5517 @dfm
@David-McKenna — Thanks! Can you update the metadata for the Zenodo deposit (there should be an "edit" button on the top right corner of that page) so that the title and author list match the paper?
