### 2024-04-22 #5517 @dfm
@pritchardn, @plaplant — Many thanks for your reviews here! And also thanks to @plaplant for getting the review started as editor. JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!
